requirements.txt
git+https://github.com/mts7/mts-logger@v0.2.4
flask
pytest
pytest-mock
python-dotenv
uwsgi
